requirements-dev.txt
-r requirements.txt
pytest
pytest-django
pytest-asyncio
pytest-cov
pytest-mock
httpx
ruff
locust
